% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUDUSD}
\alias{AUDUSD}
\title{Prices of AUD/USD FX.}
\format{A xts, zoo with 499 rows and 1 variable:
\describe{
  \item{AUD.USD}{price, in US dollars}
  ...
}}
\source{
\url{https://finance.yahoo.com/quote/AUDUSD=X?p=AUDUSD=X}
}
\usage{
AUDUSD
}
\description{
A time-series containing the 499 prices points of AUD/USD FX.
}
\keyword{datasets}
