% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentedDickeyFullerTest.R
\name{AugmentedDickeyFullerTest}
\alias{AugmentedDickeyFullerTest}
\title{Augmented Dickey Fuller Test}
\usage{
AugmentedDickeyFullerTest(y, type = "drift", lags = 1)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{type}{Test type, either "none", "drift" or "trend".}

\item{lags}{Number of lags for endogenous variable to be included.}
}
\value{
A \code{list} of ADF test results
}
\description{
ADF test is a unit-root test for determining whether the spread is cointegrated
}
\details{
If type is set to "none" neither an intercept nor a trend is included in the test regression.
If it is set to "drift" an intercept is added and if it is set to "trend" both an intercept and a trend is added.
}
\examples{
data(Raotbl3)
attach(Raotbl3)
lc.df <- AugmentedDickeyFullerTest(y=lc, lags=3, type='trend')
lc.df$res

}
