% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUSYC}
\alias{AUSYC}
\title{Australian Bonds.}
\format{A xts, zoo matrix with 5460 points of Australian Bonds from 1995-01-03 to 2016-08-17:
\describe{
  \item{AUS2Y}{2 years AUS bonds}
  ...
}}
\usage{
AUSYC
}
\description{
2, 3, 5 and 10 years Australian bonds.
}
\keyword{datasets}
