% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BollingerBands.R
\name{BollingerBands}
\alias{BollingerBands}
\title{Calculate Bollinger Bands}
\usage{
BollingerBands(y, lookback = 20, std = 2)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{lookback}{number of perioed to look back for testing}

\item{std}{number of standard deviations to calculate BB bands}
}
\value{
A \code{data.frame} with raw time-series and BB bands
}
\description{
Bollinger bands are calculated based on moving averages and moving standard deviations of time-series.
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
BollingerBands(AUDUSD/CADUSD, lookback = 20, std = 2)

}
