% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CADUSD}
\alias{CADUSD}
\title{Prices of CAD/USD FX.}
\format{A xts, zoo with 499 rows and 1 variable:
\describe{
  \item{CAD.USD}{price, in US dollars}
  ...
}}
\source{
\url{https://finance.yahoo.com/quote/CADUSD=X?p=CADUSD=X}
}
\usage{
CADUSD
}
\description{
A time-series containing the 499 prices points of CAD/USD FX.
}
\keyword{datasets}
