% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CANYC}
\alias{CANYC}
\title{Canadian Bonds.}
\format{A xts, zoo matrix with 3002 points of Canadian Bonds from 2004-08-20 to 2016-08-18:
\describe{
  \item{CAN2Y}{2 years CAN bonds}
  ...
}}
\usage{
CANYC
}
\description{
2, 3, 5 and 10 years Canadian bonds.
}
\keyword{datasets}
