% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{CoInterestRateParity}
\alias{CoInterestRateParity}
\title{Calculate Forward FX}
\usage{
CoInterestRateParity(If, Id, SFX)
}
\arguments{
\item{If}{foreign Interest Rates}

\item{Id}{domestic Interest Rates}

\item{SFX}{Spot FX prices}
}
\value{
A \code{list} of forward FX and their growth rates
}
\description{
Calculate forward FX by using Covered Interest Rate Parity formula
}
