% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{computeHoldingsPct}
\alias{computeHoldingsPct}
\title{Compute the Holdings Percentages}
\usage{
computeHoldingsPct(y.shares, x.shares, y.price, x.price)
}
\arguments{
\item{y.shares}{y shares}

\item{x.shares}{x shares}

\item{y.price}{y price}

\item{x.price}{x price}
}
\value{
A \code{vector} of percentages of x and y to be targeted
}
\description{
Compute the Holdings Percentages based on long/short shares and prices
}
