% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationTest.R
\name{CorrelationTest}
\alias{CorrelationTest}
\title{Correlation Tests}
\usage{
CorrelationTest(x, y, method = "pearson")
}
\arguments{
\item{x, }{y numeric vectors of data values.}

\item{method}{a character string naming which test should be applied. ("pearson","kendall","spearman")}
}
\value{
A \code{list} of test results and diagram
}
\description{
Tests if two series are correlated.
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
cor = CorrelationTest(AUDUSD, CADUSD)
cor

}
