% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{createSummarySheet}
\alias{createSummarySheet}
\title{Create a Summary Sheet for Backtesting}
\usage{
createSummarySheet(prices, initial.capital)
}
\arguments{
\item{initial.capital}{initial capital}

\item{y}{y series}

\item{x}{x series}
}
\value{
A \code{data.frame}
}
\description{
Create a Summary Sheet for Backtesting to record all trading activities across the time
}
