% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconomicCalendarAnalysis.R
\name{drawEconomicCalendar}
\alias{drawEconomicCalendar}
\title{Draw Economic Calendar Plots}
\usage{
drawEconomicCalendar(INSTRUMENTS = "AUD_USD", font_size = 10)
}
\arguments{
\item{font_size}{size of font}

\item{INSTRUMENT}{instrument of OANDA}
}
\description{
Detect the impacts of Economic Calendar on Forex market
}
\examples{
a = drawEconomicCalendar("AUD_USD")

}
