% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fillMissingData}
\alias{fillMissingData}
\title{Fill Missing Data}
\usage{
fillMissingData(v)
}
\arguments{
\item{v}{a vector of values}
}
\value{
A \code{vector} of filled values
}
\description{
Fill missing data according to the last value
}
\examples{
v = c(1, rep(NA, 10), 2)
fillMissingData(v)

}
