% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationTest.R
\name{findCorLags}
\alias{findCorLags}
\title{Find Most Correlated Lags}
\usage{
findCorLags(x, y, mlag = 60)
}
\arguments{
\item{x}{main series}

\item{y}{secondary series which will be applied to get lags}

\item{mlag}{maximum lags to test}
}
\value{
A \code{list} of all lags and their correlations
}
\description{
Search the number of lags with highest correlations between x and y
}
