% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{finIndicatorCommon}
\alias{finIndicatorCommon}
\title{Generate Financial Indicator (Common)}
\usage{
finIndicatorCommon(x)
}
\arguments{
\item{x}{prices data}
}
\description{
Generate Financial Indicator (Common)
}
