% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{finIndicatorHLC}
\alias{finIndicatorHLC}
\title{Generate Financial Indicator (HLC)}
\usage{
finIndicatorHLC(x)
}
\arguments{
\item{x}{prices data}
}
\description{
Generate Financial Indicator (HLC)
}
