% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{finIndicatorHLCVol}
\alias{finIndicatorHLCVol}
\title{Generate Financial Indicator (HLC + Vol)}
\usage{
finIndicatorHLCVol(x, v)
}
\arguments{
\item{x}{prices data}

\item{v}{volume}
}
\description{
Generate Financial Indicator (HLC + Vol)
}
