% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleStickPatterns.R
\name{generateCandleStickPatterns}
\alias{generateCandleStickPatterns}
\title{Generate Japanese Candlesticks Patterns}
\usage{
generateCandleStickPatterns(x)
}
\arguments{
\item{x}{TS}
}
\description{
Generate Japanese Candlesticks Patterns
}
\examples{
load('~/Common/audcad_intraday_m15/model/Data_AUDUSD_1487485108.RData')
generateCandleStickPatterns(prices.ib)

}
