% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getPriceRatio}
\alias{getPriceRatio}
\title{Get Price Ratio / Spreads between a Pair of Prices}
\usage{
getPriceRatio(y, x, log = TRUE, spread = FALSE)
}
\arguments{
\item{y}{y series}

\item{x}{x series}

\item{log}{log scaled or not}

\item{spread}{return spread instead of price ratio}
}
\value{
A \code{vector} of price ratio OR spreads
}
\description{
Get Price Ratio / Spreads between a Pair of Prices
}
