% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getSlope}
\alias{getSlope}
\title{Get the slope of a vector of numbers}
\usage{
getSlope(y)
}
\arguments{
\item{y}{a vector of numbers}
}
\value{
A \code{numer} of slope value
}
\description{
Get the slope of a vector of numbers
}
