% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHedgeRatio.R
\name{HedgeRatioOLS}
\alias{HedgeRatioOLS}
\title{Hedge Ratio OLS}
\usage{
HedgeRatioOLS(y, x, add_const = TRUE)
}
\arguments{
\item{y}{long series}

\item{x}{short series}

\item{add_const}{keep constant (risk free profits) or not}
}
\value{
A \code{list} of hedge ratio (beta, risk) and alpha (risk free profits)
}
\description{
Calculate Hedge Ratio by using OLS
}
\details{
alpha < 0 too little returns based on the risks

alpha = 0 adequate returns for the risks taken

alpha > 0 excess returns to the assumed risks

beta correlated relative volatile
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
HedgeRatioOLS(AUDUSD, CADUSD)

}
