% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurstExponentTest.R
\name{HurstExponentTest}
\alias{HurstExponentTest}
\title{Hurst Exponent Test}
\usage{
HurstExponentTest(y, lookback)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{lookback}{number of perioed to look back for testing}
}
\value{
A \code{list} of hurst test results
}
\description{
Effectively this returns a value between 0 and 1 that tells you whether a time-series is trending or mean-reverting.
The closer the value is to 0.5 means the more "random" the time-series has behaved historically.
Values below 0.5 imply the time-series is mean-reverting, and above 0.5 imply trending.
The closer the value is to 0 implies greater levels of mean-reversion.
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
HurstExponentTest(AUDUSD/CADUSD, 20)

}
