% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{indexation}
\alias{indexation}
\title{Get an Index of Current T-series}
\usage{
indexation(x)
}
\arguments{
\item{x}{time series}
}
\value{
A \code{vector} of indexed time-series
}
\description{
Divide each column by the first non-NA value to get an index of 1 for input time-series
}
