% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohansenCointegrationTest.R
\name{JohansenCointegrationTest}
\alias{JohansenCointegrationTest}
\title{Johansen Cointegration Test}
\usage{
JohansenCointegrationTest(x, type = "trace", ecdet = "none", K = 2)
}
\arguments{
\item{x}{Data matrix to be investigated for cointegration.}

\item{type}{The test to be conducted, either ‘eigen’ or ‘trace’.}

\item{ecdet}{Character, ‘none’ for no intercept in cointegration, ‘const’ for constant term in cointegration and ‘trend’ for trend variable in cointegration.}

\item{K}{The lag order of the series (levels) in the VAR.}
}
\value{
A \code{list} of Johansen Cointegration test results
}
\description{
Conducts the Johansen procedure on a given data set.
The "trace" or "eigen" statistics are reported and the matrix of eigenvectors as well as the loading matrix.
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
res <- JohansenCointegrationTest(merge(AUDUSD, CADUSD), type = "trace", ecdet = "none", K = 2)
res

}
