% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{make_order_pct}
\alias{make_order_pct}
\title{Make Orders based on Portfolio Percent}
\usage{
make_order_pct(initial.capital, capital, holdings, prices, current.shares, pct,
  brokerage = 0.001)
}
\arguments{
\item{initial.capital}{total capital to invest}

\item{capital}{current capital in hand}

\item{holdings}{current holdings in dollar values for symbol}

\item{prices}{current prices of symbol}

\item{current.shares}{current number of units of symbol}

\item{pct}{target percent of total portfolio to invest}

\item{brokerage}{brokerage fees in percentage}
}
\value{
A \code{list} of updated values for current status of the symbol
}
\description{
Make orders based on portfolio percentages
}
