% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{modelBondIRFeatures}
\alias{modelBondIRFeatures}
\title{Create Bonds and Interest Rates Features}
\usage{
modelBondIRFeatures(bond1, bond2)
}
\arguments{
\item{bond1}{main bonds}

\item{bond2}{secondary bonds}
}
\description{
Create Bonds and Interest Rates Features
}
