% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{modelCandleFeatures}
\alias{modelCandleFeatures}
\title{Consecutive Candle Analysis}
\usage{
modelCandleFeatures(data)
}
\arguments{
\item{data}{raw data of prices}
}
\description{
Consecutive Candle Analysis
}
