% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{modelMultivariableFeatures}
\alias{modelMultivariableFeatures}
\title{For loop for generating Univariate Features}
\usage{
modelMultivariableFeatures(dat)
}
\arguments{
\item{dat}{multivariate time-series}
}
\description{
For loop for generating Univariate Features
}
