% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{modelPricingFeatures}
\alias{modelPricingFeatures}
\title{Generate model data based on prices}
\usage{
modelPricingFeatures(dat, test.p = 63)
}
\arguments{
\item{dat}{raw data of prices}

\item{test.p}{lookback periods}
}
\description{
Generate model data based on prices
}
