% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentumTest.R
\name{momentum.Crossover}
\alias{momentum.Crossover}
\title{Use MA Crossovers to create different momentum indicators}
\usage{
momentum.Crossover(y)
}
\arguments{
\item{y}{Price series that is coercible to xts or matrix.}
}
\description{
Use 10 moving averages (short to long) to calculate momentum indicators.
}
\details{
Hamming distance indicator: rankings from 1 to 10 to indicate the strength of a momentum

Spearman cor indicator: a number from -1 to 1 to tell the direction of current momentum. (high means short and low means long)

Thickness: thickness/difference of 10 moving averages, higher thickness means higher volatilities or stronger momentum
}
\examples{
y = AUDUSD
momCrossover <- momentum.Crossover(y)

}
\seealso{
\link{momentum.RSI}

\link{momentum.MACD}
}
