% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentumTest.R
\name{momentum.MACD}
\alias{momentum.MACD}
\title{Use MACD to create a momentum indicator}
\usage{
momentum.MACD(y, nFast = 12, nSlow = 26, nSig = 9, nEMA = 20)
}
\arguments{
\item{y}{Price series that is coercible to xts or matrix}

\item{nFast}{Number of periods for fast moving average.}

\item{nSlow}{Number of periods for slow moving average.}

\item{nSig}{Number of periods for signal moving average.}

\item{nEMA}{Number of periods for exponentially-weighted moving average.}
}
\description{
Use MACD (moving average convergence divergence) to calculate a momentum indicator.
}
\details{
When price is going over EMA(20) and MACD is going over 0 then go long

When price is going below EMA(20) and MACD is going below 0 then go short
}
\examples{
y = AUDUSD
momMACD <- momentum.MACD(y)

}
\seealso{
\link{momentum.RSI}

\link{momentum.Crossover}
}
