% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentumTest.R
\name{momentum.RSI}
\alias{momentum.RSI}
\title{Use RSI to calculate a momentum indicator}
\usage{
momentum.RSI(y, n)
}
\arguments{
\item{y}{Price series that is coercible to xts or matrix}

\item{n}{Number of periods for moving averages.}
}
\description{
Use RSI (Relative Strength Index) to calculate a momentum indicator
}
\details{
The RSI calculation is RSI = 100 - 100 / ( 1 + RS ), where RS is the smoothed ratio of 'average' gains over 'average' losses. The 'averages' aren't true averages, since they're divided by the value of n and not the number of periods in which there are gains/losses.

RSI > 70 means the asset is overvalued and overbought (go short)

RSI < 30 means the asset is undervalued and underbought (go long)
}
\examples{
y = AUDUSD
momRSI <- momentum.RSI(y)

}
\seealso{
\link{momentum.MACD}

\link{momentum.Crossover}
}
