% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrnsteinUhlenbeckHalfLife.R
\name{OrnsteinUhlenbeckHalfLife}
\alias{OrnsteinUhlenbeckHalfLife}
\title{Calculate Half-Life for Mean-Reversion}
\usage{
OrnsteinUhlenbeckHalfLife(price.Ratio)
}
\arguments{
\item{price.Ratio}{price ratio between pairs}
}
\value{
A \code{list} of half life results
}
\description{
Computed from an Ornstein–Uhlenbeck process. This is the theoretically computed time, based on a historical window of data,
that it will take for the spread to mean-revert half of its distance after having diverged from the mean of the spread.
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
half.life <- OrnsteinUhlenbeckHalfLife(AUDUSD/CADUSD)
half.life

}
