% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepare.R
\name{prepareForexOandaData}
\alias{prepareForexOandaData}
\title{Prepare all required datasets for Oanda FX daily Modeling}
\usage{
prepareForexOandaData(oanda.count = 250, Cur1 = "USD", Cur2 = "CAD",
  oanda.granularity = "D", QuandlSymbol1 = "CAN", QuandlSymbol2 = "USA")
}
\arguments{
\item{oanda.count}{length of data to retreive}

\item{Cur1}{currency one}

\item{Cur2}{currency two}

\item{oanda.granularity}{barsize of Oanda data}

\item{QuandlSymbol1}{currency one for Quandl}

\item{QuandlSymbol2}{currency two for Quandl}
}
\value{
A \code{list} of \code{xts}
}
\description{
Prepare all required datasets for Oanda FX daily Modeling.
}
\examples{
res = prepareForexOandaData(oanda.count = 2500, Cur1 = "USD", Cur2 = "CAD", oanda.granularity = 'D', QuandlSymbol1 = 'CAN', QuandlSymbol2 = 'USA')

}
