% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepare.R
\name{prepareForexOandaPrices}
\alias{prepareForexOandaPrices}
\title{Prepare all required datasets for Oanda FX daily Modeling}
\usage{
prepareForexOandaPrices(ACCOUNT_TYPE = .oandaEnv$ACCOUNT_TYPE,
  ACCESS_TOKEN = .oandaEnv$ACCESS_TOKEN, oanda.count = 250, Cur1 = "USD",
  Cur2 = "CAD", oanda.granularity = "D")
}
\arguments{
\item{ACCOUNT_TYPE}{Account type (e.g. "real" or "paper")}

\item{ACCESS_TOKEN}{Account API Token}

\item{oanda.count}{length of data to retreive}

\item{Cur1}{currency one}

\item{Cur2}{currency two}

\item{oanda.granularity}{barsize of Oanda data}
}
\value{
A \code{list} of \code{xts}
}
\description{
Prepare all required datasets for Oanda FX daily Modeling.
}
\examples{
res = prepareForexOandaPrices(oanda.count = 2500, Cur1 = "USD", Cur2 = "CAD", oanda.granularity = 'D')

}
