% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepare.R
\name{prepareIBForexPrices}
\alias{prepareIBForexPrices}
\title{Prepare Interactive Brokers FX data}
\usage{
prepareIBForexPrices(ib.duration = "1 Y", ib.barsize = "1 day",
  Cur1 = "USD", Cur2 = "CAD", ibAcc = "paper", midOnly = TRUE)
}
\arguments{
\item{ib.duration}{length of data to retreive from Interactive Broker}

\item{ib.barsize}{the barsize of IB data}

\item{Cur1}{currency one}

\item{Cur2}{currency two}

\item{ibAcc}{account type of Interactive Brokers (live or paper)}

\item{midOnly}{only retreive mid prices}
}
\value{
A \code{list} of \code{xts}
}
\description{
Prepare Interactive Brokers FX data
}
