% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{prepareMachinelearningFeatures}
\alias{prepareMachinelearningFeatures}
\title{Prepare the features for machine learning}
\usage{
prepareMachinelearningFeatures(price.x)
}
\arguments{
\item{price.x}{prices data}
}
\description{
Prepare the features for machine learning
}
