% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{prepareMachinelearningFeaturesIG}
\alias{prepareMachinelearningFeaturesIG}
\title{Prepare the non-volume features for machine learning}
\usage{
prepareMachinelearningFeaturesIG(price.x)
}
\arguments{
\item{price.x}{prices data}
}
\description{
Prepare the non-volume features for machine learning
}
