% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleUniverse}
\alias{SampleUniverse}
\title{Sample Data for Modeling}
\format{A xts, zoo matrix with 618 points from 2014-1-02 to 2016-07-22:
\describe{
  \item{price.ratio}{price ratios between pairs}
  ...
}}
\usage{
SampleUniverse
}
\description{
All required features including AUDUSD, CADUSD and Bonds Ratios for modeling
}
\keyword{datasets}
