% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchCointegrationPairs.R
\name{searchCointegratedPairsOanda}
\alias{searchCointegratedPairsOanda}
\title{Search Cointegrated FX Pairs via Oanda API}
\usage{
searchCointegratedPairsOanda(total_points = 500, testPeriod = 0.25,
  INSTRUMENTS, dir = "~/analytics/common/cointegration/")
}
\arguments{
\item{total_points}{Total data points to lookback (e.g. days)}

\item{testPeriod}{proportion of data points to be testing set}

\item{INSTRUMENTS}{a list of available FX instruments}

\item{dir}{directory to save the output}
}
\description{
Search Cointegrated FX Pairs via Oanda API
}
\examples{
INSTRUMENTS = readRDS('~/analytics/common/OANDA_FX_INSTRUMENTS.rds')
res = searchCointegratedPairsOanda(total_points = 500, testPeriod = 0.25, INSTRUMENTS = INSTRUMENTS, dir = "~/analytics/common/cointegration/")
res$res
res$path

}
\seealso{
\link{JohansenCointegrationTest}

\link{AugmentedDickeyFullerTest}
}
