% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setDataTable}
\alias{setDataTable}
\title{Convert Summary sheet into data.table}
\usage{
setDataTable(dt)
}
\arguments{
\item{dt}{summary trading sheet}
}
\value{
A \code{data.table}
}
\description{
Convert Summary sheet into data.table
}
