% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABTest.R
\name{setupNewStrategy}
\alias{setupNewStrategy}
\title{Setup New Strategy Directory}
\usage{
setupNewStrategy(strategyName, strategyPath)
}
\arguments{
\item{strategyName}{Strategy Name}

\item{strategyPath}{Path of directory}
}
\description{
Setup New Strategy Directory
}
\examples{
setupNewStrategy("strategy1", "./Common/")

}
