% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Prices of S&P 500.}
\format{A xts, zoo matrix with 500 stocks in S&P from 2010-01-01 to 2014-01-31:
\describe{
  \item{AAPL}{price, in US dollars}
  ...
}}
\usage{
sp500
}
\description{
Prices of S&P 500 from 2010-01-01 to 2014-01-31.
}
\keyword{datasets}
