% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startAlgoTradingOanda.R
\name{startAlgoTrading}
\alias{startAlgoTrading}
\title{Main functions for Oanda Algo Systems}
\usage{
startAlgoTrading(configs)
}
\arguments{
\item{configs}{a list() of configurations required by trading systems}
}
\description{
Main functions for Oanda Algo Systems
}
\examples{
rm(list=ls());
configs = list(
order.book.file = '~/Common/audcad_intraday_m5/orderbook/ivanliu_AUDCAD_Pairs_M5_003.csv'
,log.file = file(paste0('~/Common/audcad_intraday_m5/log/ivanliu_AUDCAD_Pairs_M5_003_', as.Date(Sys.time()), '.Rout'), open="wt")
,plotPath = "~/Common/audcad_intraday_m5/plots/audcad_intraday_m5_"
,strategyPath = '~/Common/audcad_intraday_m5/'
,pair1 = c('AUD', 'USD') # Later use
,pair2 = c('USD', 'CAD') # Later use
,initAssets = 1000000
,posiRate = 0.05
,threshold = 1.8
,mean.threshold = 0.2
,moment.threshold = 0.75 # effect * direction
,stoploss = -0.10
,ACCOUNT_ID = ' '
,train.length = 1000
,min.lookback = 48
,max.lookback = 240
,sd.n = 3
,freq = 'M5'
)

rm(list=ls());
configs = list(
    order.book.file = '~/Common/audcad_interday_d1/orderbook/ivanliu_AUDCAD_Pairs_Daily_001.csv'
    ,log.file = file(paste0('~/Common/audcad_interday_d1/log/ivanliu_AUDCAD_Pairs_Daily_001_', as.Date(Sys.time()), '.Rout'), open = 'wt')
    ,plotPath = "~/Common/audcad_interday_d1/plots/audcad_intraday_d1_"
    ,strategyPath = '~/Common/audcad_interday_d1/'
    ,pair1 = c('AUD', 'USD') # Later use
    ,pair2 = c('USD', 'CAD') # Later use
    ,initAssets = 1000000
    ,posiRate = 0.1
    ,threshold = 1.8
    ,mean.threshold = 0.2
    ,moment.threshold = 0.75 # effect * direction
    ,stoploss = -0.10
    ,ACCOUNT_ID = '101-011-4686012-001'
    ,train.length = 504
    ,min.lookback = 21
    ,max.lookback = 63
    ,sd.n = 3
    ,freq = 'D'
    )

rm(list=ls());
configs = list(
    order.book.file = '~/Common/audcad_intraday_h1/orderbook/ivanliu_AUDCAD_Pairs_Hourly_002.csv'
    ,log.file = file(paste0('~/Common/audcad_intraday_h1/log/ivanliu_AUDCAD_Pairs_Hourly_002_', as.Date(Sys.time()), '.Rout'), open = 'wt')
    ,plotPath = "~/Common/audcad_intraday_h1/plots/audcad_intraday_h1_"
    ,strategyPath = '~/Common/audcad_intraday_h1/'
    ,pair1 = c('AUD', 'USD') # Later use
    ,pair2 = c('USD', 'CAD') # Later use
    ,initAssets = 1000000
    ,posiRate = 0.1
    ,threshold = 1.8
    ,mean.threshold = 0.49
    ,moment.threshold = 0.75 # effect * direction
    ,stoploss = -0.10
    ,ACCOUNT_ID = '101-011-4686012-003'
    ,train.length = 1000
    ,min.lookback = 48
    ,max.lookback = 240
    ,sd.n = 3
    ,freq = 'H1'
    )

startAlgoTrading(configs)
}
