% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABTest.R
\name{storeOandaPricing}
\alias{storeOandaPricing}
\title{Store the historical Pricing points from Oanda}
\usage{
storeOandaPricing(Cur1 = "AUD", Cur2 = "CAD",
  ACCOUNT_TYPE = .oandaEnv$ACCOUNT_TYPE,
  ACCESS_TOKEN = .oandaEnv$ACCESS_TOKEN,
  dir = "~/Common/models/AUDCAD_Oanda.RData")
}
\arguments{
\item{Cur1}{Currency 1}

\item{Cur2}{Currency 2}

\item{ACCOUNT_TYPE}{Account type (e.g. "real" or "paper")}

\item{ACCESS_TOKEN}{Account API Token}

\item{dir}{Data file directory}
}
\description{
Store the historical Pricing points from Oanda
}
\examples{
storeOandaPricing(Cur1 = 'AUD', Cur2 = 'CAD', ACCOUNT_TYPE = .oandaEnv$ACCOUNT_TYPE, ACCESS_TOKEN = .oandaEnv$ACCESS_TOKEN, dir = "~/Common/models/AUDCAD_Oanda.RData")

}
