% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterSentimentAnalysis.R
\name{twitterClusterGraph}
\alias{twitterClusterGraph}
\title{Cluster Graph of Tweets on certain Topic}
\usage{
twitterClusterGraph(tweets, minDocFreq = 0.9, groups = 6,
  title = "AUD/USD", dir = "~/analytics/common/")
}
\arguments{
\item{tweets}{a list of corpus and data.frame of tweets retreived from \code{tweet_corpus_chunck()} or \code{tweet_corpus{}}}

\item{minDocFreq}{minimum frequenciesof each term in percent in all documents}

\item{groups}{number of clusters}

\item{title}{title to be added to the results plots}

\item{dir}{directory to save the plots}
}
\description{
Which words are closer to each other? Cluster Graph of Tweets on certain Topic.
}
\details{
Please use \code{RQuantAPI} for twitter data retrieval
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus_chunck(search = "aud+usd", n = 50000, chunck_size = 7, total_size = 90, until = Sys.Date())
saveRDS(tweets, '~/analytics/Workspace/Tweets.RDS')
tweets = readRDS('~/analytics/Workspace/Tweets.RDS')
twitterClusterGraph(tweets, minDocFreq = 0.95, groups = 6, title ="AUD/USD")

}
