% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterSentimentAnalysis.R
\name{twitterEmotionalValence}
\alias{twitterEmotionalValence}
\title{Emotional valence of Tweets on certain Topic}
\usage{
twitterEmotionalValence(tweets, title = "AUD/USD",
  dir = "~/analytics/common/", fontsize = 10)
}
\arguments{
\item{tweets}{a list of corpus and data.frame of tweets retreived from \code{tweet_corpus_chunck()} or \code{tweet_corpus{}}}

\item{title}{title to be added to the results plots}

\item{dir}{directory to save the plots}

\item{fontsize}{font size}
}
\description{
When are people's attitudes? Emotional valence of Tweets on certain Topic.
}
\details{
Please use \code{RQuantAPI} for twitter data retrieval
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus_chunck(search = "aud+usd", n = 50000, chunck_size = 7, total_size = 90, until = Sys.Date())
saveRDS(tweets, '~/analytics/Workspace/Tweets.RDS')
tweets = readRDS('~/analytics/Workspace/Tweets.RDS')
emotion = twitterEmotionalValence(tweets, "AUD/USD")
emotion$orig$text_raw[which.max(emotion$orig$emotionalValence)]
emotion$orig$text_raw[which.min(emotion$orig$emotionalValence)]
emotion$emotionDailyPlot
emotion$emotionPlotDailyBox
emotion$emotionPlot
emotion$emotionPlotViolin
emotion$wc_filename

}
