% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USAYC}
\alias{USAYC}
\title{Canadian Bonds.}
\format{A xts, zoo matrix with 6653 points of USA Bonds from 1990-1-02 to 2016-08-02:
\describe{
  \item{USA2Y}{2 years USA bonds}
  ...
}}
\usage{
USAYC
}
\description{
2, 3, 5 and 10 years USA bonds.
}
\keyword{datasets}
