% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{volatilityTraining}
\alias{volatilityTraining}
\title{Volatility Machine Learning Models}
\usage{
volatilityTraining(ACCOUNT_TYPE = .oandaEnv$ACCOUNT_TYPE,
  ACCESS_TOKEN = .oandaEnv$ACCESS_TOKEN, Cur1, Cur2, nxt.n, neval = 550,
  ntrain = "1 Y", ib.size = "1 hour", oa.size = "H1")
}
\arguments{
\item{ACCOUNT_TYPE}{Account type (e.g. "real" or "paper")}

\item{ACCESS_TOKEN}{Account API Token}

\item{Cur1}{currency 1}

\item{Cur2}{currency 2}

\item{nxt.n}{number of lags to predict}

\item{neval}{number of evaluations}

\item{ntrain}{number of obs for training set}

\item{ib.size}{1 hour}

\item{oa.size}{H1}
}
\value{
A \code{list} of Models
}
\description{
Volatility Machine Learning Models
}
\examples{
volatilityModel = volatilityTraining('AUD', 'USD', 3, 550, '1 Y')

}
