% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{zscores.ma}
\alias{zscores.ma}
\title{Zscore based Moving Average}
\usage{
zscores.ma(t, long = 60, short = 10)
}
\arguments{
\item{t}{a time-series to be normalized}

\item{long}{days of long-term moving averages}

\item{short}{days of short-term moving averages}
}
\value{
A normalized time-series
}
\description{
Using moving averages to compute the z-score of the ratio at each given time.
This will tell us how extreme the ratio is and whether it's a good idea to enter a position at this time.
}
\examples{
library(quantmod)
getFX("AUD/USD")
zscores.ma(AUDUSD, 60, 10)

}
