% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{zscores}
\alias{zscores}
\title{Zscore}
\usage{
zscores(t)
}
\arguments{
\item{t}{a time-series to be normalized}
}
\value{
A normalized time-series
}
\description{
Normalize time series
}
\examples{
library(quantmod)
getFX("AUD/USD")
zscores(AUDUSD)

}
